requirements.txt
evdev==1.6.1
python-uinput==0.11.2
python-xlib==0.33
six==1.16.0
